[HMA-4097] Add savings total to update

+++ app/uk/gov/hmrc/mobilehelptosave/domain/SavingsUpdateResponse.scala
@@ -31,7 +31,7 @@ object SavingsUpdateResponse {
 }
 
 case class SavingsUpdate(
-  savedInPeriod:            Option[Double],
+  savedInPeriod:            Option[BigDecimal],
   monthsSaved:              Option[Int],
   goalsReached:             Option[GoalsReached],
   amountEarnedTowardsBonus: Option[Double])
